Introduce archives.h/cpp
