+++ toolsrc/vcpkglib/vcpkglib.vcxproj
@@ -137,6 +137,7 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="..\include\pch.h" />
+    <ClInclude Include="..\include\vcpkg\base\archives.h" />
     <ClInclude Include="..\include\vcpkg\base\checks.h" />
     <ClInclude Include="..\include\vcpkg\base\chrono.h" />
     <ClInclude Include="..\include\vcpkg\base\cofffilereader.h" />
@@ -193,6 +194,7 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
     </ClCompile>
+    <ClCompile Include="..\src\vcpkg\base\archives.cpp" />
     <ClCompile Include="..\src\vcpkg\base\checks.cpp" />
     <ClCompile Include="..\src\vcpkg\base\chrono.cpp" />
     <ClCompile Include="..\src\vcpkg\base\cofffilereader.cpp" />
